requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
PyJWT==2.4.0
pymongo==4.2.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.2.2
